reorg patches in CustomFeatures
