+++ SolastaCommunityExpansion/Patches/CustomFeatures/RitualAndAutoPreparedSpells/RulesetCharacterPatcher.cs
@@ -4,59 +4,8 @@ using HarmonyLib;
 using SolastaCommunityExpansion.CustomFeatureDefinitions;
 using SolastaCommunityExpansion.Models;
 
-namespace SolastaCommunityExpansion.Patches.CustomFeatures
+namespace SolastaCommunityExpansion.Patches.CustomFeatures.RitualAndAutoPreparedSpells
 {
-    [HarmonyPatch(typeof(RulesetCharacter), "UsePower")]
-    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
-    internal static class RulesetCharacter_UsePower
-    {
-        public static void Postfix(RulesetCharacter __instance, RulesetUsablePower usablePower)
-        {
-            __instance.UpdateUsageForPowerPool(usablePower, usablePower.PowerDefinition.CostPerUse);
-        }
-    }
-
-    [HarmonyPatch(typeof(RulesetCharacter), "RepayPowerUse")]
-    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
-    internal static class RulesetCharacter_RepayPowerUse
-    {
-        public static void Postfix(RulesetCharacter __instance, RulesetUsablePower usablePower)
-        {
-            __instance.UpdateUsageForPowerPool(usablePower, -usablePower.PowerDefinition.CostPerUse);
-        }
-    }
-
-    [HarmonyPatch(typeof(RulesetCharacter), "GrantPowers")]
-    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
-    internal static class RulesetCharacter_GrantPowers
-    {
-        public static void Postfix(RulesetCharacter __instance)
-        {
-            CustomFeaturesContext.RechargeLinkedPowers(__instance, RuleDefinitions.RestType.LongRest);
-        }
-    }
-
-    [HarmonyPatch(typeof(RulesetCharacter), "ApplyRest")]
-    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
-    internal static class RulesetCharacter_ApplyRest
-    {
-        internal static void Postfix(
-            RulesetCharacter __instance, RuleDefinitions.RestType restType, bool simulate)
-        {
-            if (!simulate)
-            {
-                CustomFeaturesContext.RechargeLinkedPowers(__instance, restType);
-            }
-
-            // The player isn't recharging the shared pool features, just the pool.
-            // Hide the features that use the pool from the UI.
-            foreach (FeatureDefinition feature in __instance.RecoveredFeatures.Where(f => f is IPowerSharedPool).ToArray())
-            {
-                __instance.RecoveredFeatures.Remove(feature);
-            }
-        }
-    }
-
     [HarmonyPatch(typeof(RulesetCharacter), "ComputeAutopreparedSpells")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
     internal static class RulesetCharacter_ComputeAutopreparedSpells
